[refactor]コードの可読性と保守性を向上
