+++ constants.js
@@ -1,0 +1,10 @@
+// constants.js
+export const CHAR_MAP = "AIQYgow5BJRZhpx6CKSaiqy7DLTbjrz8EMUcks19FNVdlt2!GOWemu3?HPXfnv4/";
+export const MAP1_EXPANSION = "eABCDEFGHI";
+export const MAP2_EXPANSION = "pJKLMNOPQR";
+export const CARD_IMAGE_PATH = './cards/';
+export const DOUBLE_TAP_THRESHOLD = 300;
+export const LONG_PRESS_DELAY = 500;
+export const AUTH_COOKIE_NAME = 'kamitsubaki-auth';
+export const AUTH_EXPIRY_DAYS = 30;
+export const CORRECT_PASSWORD_HASH = '9d2a18cc82a3b5bf3d932c1f562f7043066b3fb777d4e00b4dec71de2b8bc5b5'; // SHA-256 hash of password
