+++ index.html
@@ -30,20 +30,20 @@
                 <button id="dual-deck-mode-btn" class="mode-btn">2デッキ対戦モード</button>
             </div>
             <div id="single-deck-input" class="deck-input-section">
-                <p>デッキリストを / 区切りで入力してください</p>
-                <textarea id="deck-string" placeholder="例: AA-01/AA-01/AA-02/..."></textarea>
+                <p>デッキコードをKCG形式かスラッシュ区切りで入力してください</p>
+                <textarea id="deck-string" placeholder="例: KCG-8DLMyRZud... or AA-1/AA-6/AA-6/..."></textarea>
                 <button id="clear-deck-btn" class="clear-btn">クリア</button>
             </div>
 
             <div id="dual-deck-input" class="deck-input-section" style="display: none;">
                 <div class="deck-input-group">
                     <h3>プレイヤー1のデッキ</h3>
-                    <textarea id="deck-string-p1" placeholder="例: AA-01/AA-01/AA-02/..."></textarea>
+                    <textarea id="deck-string-p1" placeholder="例: KCG-8DLMyRZud... or AA-1/AA-6/AA-6/..."></textarea>
                     <button id="clear-deck-p1-btn" class="clear-btn">クリア</button>
                 </div>
                 <div class="deck-input-group">
                     <h3>プレイヤー2のデッキ</h3>
-                    <textarea id="deck-string-p2" placeholder="例: BA-01/BA-01/BA-02/..."></textarea>
+                    <textarea id="deck-string-p2" placeholder="例: KCG-8DLMyRZud... or AA-1/AA-6/AA-6/..."></textarea>
                     <button id="clear-deck-p2-btn" class="clear-btn">クリア</button>
                 </div>
             </div>
@@ -220,7 +220,7 @@
         </div>
     </div>
 
-    <script src="script.js"></script>
+    <script type="module" src="main.js"></script>
     <script>
         if ('serviceWorker' in navigator) {
             window.addEventListener('load', () => {
